requirements.txt
amqp==2.5.0
anyjson==0.3.3
billiard==3.6.0.0
celery==4.3.0
certifi==2019.3.9
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
decorator==4.4.2
Django==2.2.4
django-celery-beat==1.4.0
django-celery-monitor==1.1.2
django-celery-results==1.0.4
django-coreapi==2.3.0
django-cors-headers==3.6.0
django-filter==2.2.0
django-rest-swagger==2.2.0
django-timezone-field==3.0
djangorestframework==3.9.2
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
kombu==4.6.3
Markdown==3.1.1
MarkupSafe==1.1.1
openapi-codec==1.3.2
pydash==4.9.2
python-crontab==2.3.8
python-dateutil==2.8.0
pytz==2019.2
requests==2.22.0
self==2020.12.3
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3
vine==1.3.0
